+++ backend/db/models/spot.js
@@ -51,7 +51,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     lat: {
-      type: DataTypes.DECIMAL(10,7),
+      type: DataTypes.FLOAT(10,7),
       allowNull: false,
       validate: {
         min: -90,
@@ -59,7 +59,7 @@ module.exports = (sequelize, DataTypes) => {
       }
     },
     lng: {
-      type: DataTypes.DECIMAL(10,7),
+      type: DataTypes.FLOAT(10,7),
       allowNull: false,
       validate: {
         min: -180,
@@ -75,7 +75,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     price: {
-      type: DataTypes.DECIMAL,
+      type: DataTypes.FLOAT,
       allowNull: false
     }
   }, {
